+++ db/create.php
@@ -7,34 +7,18 @@ require __DIR__ . '/../vendor/autoload.php';
 MyDotenv::boot(__DIR__ . '/../.env');
 
 $appenv = $_ENV['APP_ENV'];
-$f = SqliteHelper::DbFilename();
 
-if (strtolower($appenv) != 'test') {
-    $msg = "
-==============================
-DANGER! Replacing non-test db.
-==============================
+$dbfilename = $_ENV['DB_FILENAME'];
 
-You are going to REPLACE db
-{$f},
-which may have data in it.
-
-This could be really really bad.
-
-To confirm that you want to create this db,
-please type 'confirm' at the prompt.
-
-";
-    echo $msg;
-
-    $ret = readline('type confirm here: ');
-
-    if ($ret != 'confirm') {
-        echo "\nQuitting ...\n\n";
-        return;
-    }
+if (! str_contains($dbfilename, 'test_lute.db')) {
+    $msg = "\n\n
+NOT A TESTING DB
+----------------
+Db = {$dbfilename} ... MUST contain test_lute.db\n\n";
+    throw new \Exception($msg);
 }
-
-SqliteHelper::CreateDb();
-echo "\nCreated new db {$f}\n\n";
+else {
+    SqliteHelper::CreateDb();
+    echo "\nCreated new db {$dbfilename}\n\n";
+}
 ?>