+++ src/Form/LanguageType.php
@@ -80,12 +80,21 @@ class LanguageType extends AbstractType
                     'required' => true
                   ]
             )
+            ->add('LgRightToLeft',
+                  ChoiceType::class,
+                  [ 'choices'  => [ 'No' => 0, 'Yes' => 1 ],
+                    'label' => 'Right-to-left',
+                    'help' => 'For languages such as Arabic and Hebrew',
+                    'required' => true
+                  ]
+            )
             ->add('LgParserType',
                   ChoiceType::class,
                   [ 'choices'  => [
                       'Space Delimited' => 'spacedel',
                       'Japanese Language (MeCab)' => 'japanese',
-                      'Classical Chinese' => 'classicalchinese'
+                      'Classical Chinese' => 'classicalchinese',
+                      'Turkish' => 'turkish',
                      ],
                     'label' => 'Parse as',
                     'required' => true