+++ tests/acceptance/TermUpload_Test.php
@@ -20,15 +20,18 @@ class TermUpload_Test extends AcceptanceTestBase
      */
     public function test_upload_terms_valid_file(): void
     {
+        $wait = function() { usleep(200 * 1000); };  // hack
         $this->client->request('GET', '/');
         $this->client->clickLink('Import Terms');
         $ctx = $this->getTermUploadContext();
 
         $test_file = __DIR__ . '/Fixtures/term_import_1.csv';
         $ctx->uploadFile($test_file);
+        $wait();
 
         $this->client->request('GET', '/');
         $this->client->clickLink('Terms');
+        $wait();
         $ctx = $this->getTermContext();
         $ctx->listingShouldContain(
             'two terms',
@@ -36,6 +39,7 @@ class TermUpload_Test extends AcceptanceTestBase
               '; gatos; gato; ; Spanish; ; New (1)' ]);
 
         $this->client->clickLink('gato');
+        $wait();
         $crawler = $this->client->refreshCrawler();
         $form = $crawler->selectButton('Save')->form();
         $this->assertEquals($form['term_dto[Text]']->getValue(), 'gato', 'same term found');
@@ -49,13 +53,17 @@ class TermUpload_Test extends AcceptanceTestBase
     {
         $this->client->request('GET', '/');
         $this->client->clickLink('Import Terms');
+        $wait = function() { usleep(200 * 1000); };  // hack
+        $wait();
         $ctx = $this->getTermUploadContext();
 
         $test_file = __DIR__ . '/Fixtures/term_import_variable_columns.csv';
         $ctx->uploadFile($test_file);
+        $wait();
 
         $this->client->request('GET', '/');
         $this->client->clickLink('Terms');
+        $wait();
         $ctx = $this->getTermContext();
         $ctx->listingShouldContain(
             'two terms',
@@ -90,13 +98,17 @@ class TermUpload_Test extends AcceptanceTestBase
 
         $this->client->request('GET', '/');
         $this->client->clickLink('Import Terms');
+        $wait = function() { usleep(200 * 1000); };  // hack
+        $wait();
         $ctx = $this->getTermUploadContext();
 
         $test_file = __DIR__ . '/Fixtures/term_import_issue_50_hsk.csv';
         $ctx->uploadFile($test_file);
+        $wait();
 
         $this->assertSelectorTextContains('body', 'Learning Using Texts (LUTE)');
         $this->client->clickLink('Terms');
+        $wait();
         $ctx = $this->getTermContext();
         $ctx->listingShouldContain(
             'two terms imported',